chore: Move the classes directory to lib
