+++ .gitignore
@@ -1,3 +1,8 @@
 node_modules/
 .env
 .vscode
+*.json
+*.zip
+
+!package.json
+!package-lock.json
